change extract_url.pl to urlscan
For handling many different MIME and mail encodings, extract_url.pl is
much better than urlscan. The drawback is that when used in mutt, it
parse the same URLs multiple time, some of which parsed with some junk
around the URL and make most browsers can't open it. This hinders the
user by imposing two additional task, that is the user must search the
good URL and the user must press more keys to get to the right URLs.

Because of this, I'm changing the URL extractor to urlscan, as it at
least don't give multiple URLs, and it parse better.